% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_analyze_t_test.R
\name{bio_analyze_t_test}
\alias{bio_analyze_t_test}
\title{Analyse compl\if{html}{\out{<c3>}}\if{html}{\out{<a8>}}te d'un T-Test}
\usage{
bio_analyze_t_test(
  n1,
  mean1,
  sd1,
  n2,
  mean2,
  sd2,
  test_type,
  alpha = 0.05,
  alternative = "two.sided"
)
}
\arguments{
\item{n1}{Taille de l'\if{html}{\out{<c3>}}\if{html}{\out{<a9>}}chantillon du premier groupe}

\item{mean1}{Moyenne du premier groupe}

\item{sd1}{\if{html}{\out{<c3>}}<89>cart-type du premier groupe}

\item{n2}{Taille de l'\if{html}{\out{<c3>}}\if{html}{\out{<a9>}}chantillon du deuxi\if{html}{\out{<c3>}}\if{html}{\out{<a8>}}me groupe}

\item{mean2}{Moyenne du deuxi\if{html}{\out{<c3>}}\if{html}{\out{<a8>}}me groupe}

\item{sd2}{\if{html}{\out{<c3>}}<89>cart-type du deuxi\if{html}{\out{<c3>}}\if{html}{\out{<a8>}}me groupe}

\item{test_type}{Type de test (two.sample.classique, two.sample.Welch, paired, one.sample)}

\item{alpha}{Niveau de signification (ex: 0.05)}

\item{alternative}{Hypoth\if{html}{\out{<c3>}}\if{html}{\out{<a8>}}se alternative ("two.sided", "greater", "less")}
}
\value{
Une liste contenant les r\if{html}{\out{<c3>}}\if{html}{\out{<a9>}}sultats du test, la puissance et le graphique
}
\description{
Analyse compl\if{html}{\out{<c3>}}\if{html}{\out{<a8>}}te d'un T-Test

bio_analyze_t_test(n1 = 30, mean1 = 70, sd1 = 10,
n2 = 30, mean2 = 75, sd2 = 10,
test_type = "two.sample.classique",
alpha = 0.05, alternative = "two.sided")
}
\examples{
bio_analyze_t_test(n1 = 30, mean1 = 70, sd1 = 10, n2 = 30, mean2 = 75, sd2 = 10, 
               test_type = "two.sample.classique", alpha = 0.05, alternative = "two.sided")

}
