% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{Boxplots_with_Trendline}
\alias{Boxplots_with_Trendline}
\title{Boxplots with Trendline}
\usage{
Boxplots_with_Trendline(var1, var2)
}
\arguments{
\item{var1}{One variable}

\item{var2}{Another variable}
}
\value{
a boxplot with trendlinie
}
\description{
The function takes two variables, var1 and var2, as inputs. It then creates a boxplot of var2 against var1, with a trendline added.
}
\details{
Task 2a vi.
}
