% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{calculateBivariateStats}
\alias{calculateBivariateStats}
\title{Calculate Descriptive and Inferential Statistics
for a Metric and a Dichotomous Variable}
\usage{
calculateBivariateStats(data, metricVar, dichotomousVar)
}
\arguments{
\item{data}{A dataframe containing the dataset.}

\item{metricVar}{The name of the metric variable as a string.}

\item{dichotomousVar}{The name of the dichotomous
variable as a string.}
}
\description{
This function calculates and outputs the mean, standard deviation,
and performs a t-test to assess the difference in a metric variable
across two groups defined by a dichotomous variable.
It's useful for understanding how the metric variable differs between
two groups, such as comparing the average fare paid by passengers who
survived vs. those who did not.
}
\details{
Task 2a iv.
}
