% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{chisq_test_plus_table}
\alias{chisq_test_plus_table}
\title{Bivariate descriptive statistics for categorial variables}
\usage{
chisq_test_plus_table(kat_var1, kat_var2)
}
\arguments{
\item{kat_var1}{A kategorical variable}

\item{kat_var2}{A second kategorical variable with the same length as the first}
}
\description{
Calcualates and prints the results of descriptive bivariate statistics of a
contengency table  and a chi squared test for two categorical variables.
}
\details{
Task 2a iii.
}
