% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_template.r
\name{create_template}
\alias{create_template}
\title{create a template for a roxygenizable file.}
\usage{
create_template(type = "template", file_name = ".")
}
\arguments{
\item{type}{The type of template to provide. Give the empty string ('') to
list available types.}

\item{file_name}{The name of the R code file to be created.}
}
\value{
See \code{\link{file.copy}}.
}
\description{
a simple wrapper to copy a system file to the working directory.
}
\section{Version}{
 $Id: 38a4969351d9577aac18a25e178e6a734430dec4 $
}
\examples{
create_template(file_name = 'my_r_file.r', type = 'roxygen_markdown')
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

