% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{describe_categorial}
\alias{describe_categorial}
\title{Descriptive statistics for a categorial variable}
\usage{
describe_categorial(x, name = deparse(match.call()$x), na.rm = TRUE)
}
\arguments{
\item{x}{A factor (may be ordered).}

\item{name}{The name of the factor. By default whatever was passed to x so
for describe_categorial(data$Pclass) name would default to "data$Pclass".}

\item{na.rm}{Indicates if NA values should be removed before calculation.}
}
\description{
Prints descriptive statistics for a categorial variable x and additional
ones for ordered factors. Namely the calculated statistics are the
statistical mode, empirical entropy and phi measure of dispersion and
optionally the median and quartiles.
}
\details{
Task 2a ii.
}
