% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{describe_metric}
\alias{describe_metric}
\title{Descriptive statistics for a metric variable}
\usage{
describe_metric(x)
}
\arguments{
\item{x}{A metric variable.}
}
\value{
A one row dataframe with the different statistics in columns.
}
\description{
Returns some descriptive statistics for a metric variable in form of a
dataframe for simple formatting. All NA values are removed before
calculation.
}
\details{
Task 2a i.
}
