% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{empirical_entropy}
\alias{empirical_entropy}
\title{Empirical Entropy}
\usage{
empirical_entropy(x, na.rm = TRUE)
}
\arguments{
\item{x}{A categorial variable.}

\item{na.rm}{Indicates if NA values should be removed before calculation.}
}
\value{
A floating point number between 0 and 1.
}
\description{
Computes the average empirical entropy (according to Boltzmann's formula) of
all levels. This is a measure of dispersion that returns 1 if every level
occurs with the same frequency and 0 if only one level occurs 100\% of the
time.
}
\details{
Task 2b
}
