% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{encode_variables}
\alias{encode_variables}
\title{Encode Variables}
\usage{
encode_variables(dataset)
}
\arguments{
\item{dataset}{The titanic dataset.}
}
\value{
The same dataset with the above mentioned columns as factors.
}
\description{
Turns the columns "Sex", "Survived" and "Embarked" into a
factor and "Pclass" into an ordered factor.
}
\details{
Task 1
}
