% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{extract_cabin_data}
\alias{extract_cabin_data}
\title{Extract Cabin Data}
\usage{
extract_cabin_data(dataset)
}
\arguments{
\item{dataset}{The original dataset}
}
\value{
The same dataset with the additional columns Deck and Side
}
\description{
Deck (which is the letter the cabin starts with) and Side (which is starboard
for odd and port for even cabin numbers) are extracted from Cabin and encoded
as factors.
}
\details{
Task 1

In the event that a combination of cabins is listed for each a group
of multiple people the cabins are distributed arbitrarily among the group.
Example: If there are four people in the dataset that have "A11 B12 C13" as
the cabin, these four people will be assigned "A11" "B12" "C13" "A11". This
way the number of people with a cabin can be guaranteed to be the same as in
the original dataset and the distribution of cabins should roughly match the
original data.
}
