% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{extract_salutations}
\alias{extract_salutations}
\title{Extract Salutations}
\usage{
extract_salutations(dataset)
}
\arguments{
\item{dataset}{The titanic dataset.}
}
\value{
The same dataset with the additional column "Salutation".
}
\description{
Extracts the salutations from the name column and creates a new column
with the extracted salutations in the dataset.
}
\details{
Task 1
}
