% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state(). A function to draw out points representing geographic locations of accidents.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer representing each of States of the U.S.}

\item{year}{A numeric value in the format YYYY}
}
\value{
Plots a state's layout with points, depicting the geolocation of accidents in the specific area
}
\description{
It accepts two arguments, state.num, representing an integer corresponding to each one of the 56 States
and year, as numeric value in the format YYYY
Some data from specific states are not included, specifically of (3,7,14,43,52). If such a case arises
a message is returned with the info: "no accidents to plot"
Otherwise, in case that geografic coordinates lie within specific boundaries, lon :<900 lat <90
a plot is drawn with the graphics::points() function, with each point showing the geografic location of
accident in the specific State.
Data are filtered for the specific state.num provided with dplyr::filter
}
\details{
These functions ARE NOT TO BE USED INTERACTIVELY, because each one of them encapsulates the previous one
}
\examples{

\dontrun{
fars_map_state(20,2013)}
"Plots a map of points representing accident location in the State"
\dontrun{

"It throws an error in case an invalid state number is provided
`fars_map_state(7,2014)`, which will stop execution and return the message `invalid STATE number: .`
In the case that no data of accidents are found for the combination of arguments provided
a message of `no accidents to plot` and `NULL` is returned
`fars_map_state(2,2015)`"
}

}
