% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{fars_read() Function to read in data from the working directory.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{as input serves the string of the filename, which is a compressed csv file}
}
\value{
Returns a tibble of various attributes of the data. We're only interested in the
columns" STATE, MONTH & YEAR, which provide geolocation info about accidents
}
\description{
A simple function to read in  a data file in the form of csv.bz2 using the functionality of the
readr::read_csv function. It throws an error, in case the file specified is not existent in the working directory
All test data are included in the package and can be accessed from the \verb{US National Highway Traffic Safety Administration\\'s Fatality Analysis Reporting System}.
}
\examples{
library(Documentation)
\dontrun{fars_read("accident_2013.csv.bz2")}




\dontrun{
"Throws an error in case of no existent files"
  fars_read("accident_2012.csv.bz2")
}

}
