% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years(). A function to filter data for specific years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric vector of year values, e.g. c(2013,2014,2015)}
}
\value{
A list of tibbles of the same length as the length of the input vector
}
\description{
It accepts a vector of numeric values that must correspond to year notation YYYY, e.g. 2013
It returns a list of three tibbles, each comprised of two columns, namely: MONTH & year, which themselves are numeric columns
it uses the vectorised function to lapply, which is applied on the input vector and retrieves the
necessary columns , i.e. Month & year
}
\details{
These functions ARE NOT TO BE USED INTERACTIVELY, because each one of them encapsulates the previous one
}
\examples{
fars_read_years(c(2013,2014))
\dontrun{
"Catches an error in case an invalid year is specified, without stopping the code execution
with the implementation of `tryCatch()` and returns a warning invalid year: 2012
`fars_read_years(c(2013,2012))`"
}
}
