% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years(). A function to summarize particular attributes of the data.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric vector of year values, e.g. c(2013,2014,2015)}
}
\value{
A tibble of columns "MONTH" and each one of the years specified in the input vector.
}
\description{
It accepts a vector of numeric values that must correspond to year notation YYYY, e.g. 2013
Row-binds the output of the corresponding tibbles produced from fars_read_years(), groups by year and month
and sums up for each combination. Appropriate output is rendered with use of the spread() function
of the tidyr package.
}
\details{
These functions ARE NOT TO BE USED INTERACTIVELY, because each one of them encapsulates the previous one
}
\examples{
fars_summarize_years(c(2013,2014,2015))

}
