% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{generate_preprocessed_data}
\alias{generate_preprocessed_data}
\title{Generate Preprocessed Data}
\usage{
generate_preprocessed_data()
}
\description{
Generates the preprocessed data from the dataset in "titanic.csv" by
running all the preprocessing functions and saves the cleaned dataset
into the file "titanic_cleaned.Rds".
}
\details{
Task 1
}
