% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{infer_age}
\alias{infer_age}
\title{Infer Age}
\usage{
infer_age(dataset)
}
\arguments{
\item{dataset}{The titanic dataset.}
}
\value{
The same dataset with a mutated "Age" column
}
\description{
Fills in the missing ages in the dataset using the salutations
}
\details{
Task 1
}
