% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{make_filename(). Help function.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{as numeric value in the format YYYY, e.g. 2013}
}
\value{
Returns a filename as text,e.g.:"accident_2013.csv.bz2"
}
\description{
Help function that can be used in conjunction with the other functions of the package.
It takes as input a year as_numeric and returns a text file
appropriately composed, so as to conform with the filenames of the compressed files provided
}
\examples{
make_filename(2013)

"It can also be used as a helper function, for example in reading in files with `fars_read()` "
fars_read(\link{make_filename(2013)})


}
