% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{Mosaikplot}
\alias{Mosaikplot}
\title{Mosaicplot}
\usage{
Mosaikplot(var1, var2)
}
\arguments{
\item{var1}{One variable}

\item{var2}{Another variable}
}
\value{
a mosaicplot
}
\description{
The function takes two variables, var1 and var2, as inputs. It then creates a mosaic plot using these variables
}
\details{
Task 2a vi.
}
