% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{phi_dispersion}
\alias{phi_dispersion}
\title{Phi Measure of Dispersion for Categorial Data}
\usage{
phi_dispersion(x, na.rm = TRUE)
}
\arguments{
\item{x}{A categorial variable.}

\item{na.rm}{Indicates if NA values should be removed before calculation.}
}
\value{
A floating point number between 0 and 1 (lower than entropy).
}
\description{
Computes the phi measure of dispersion which produces more intuitive numbers
between 0 and 1 but is not popular unlike entropy.
}
\details{
Task 2b
}
