\name{provide_pandoc}
\alias{provide_pandoc}
\title{install pandoc if not already installed on the system}
\usage{
provide_pandoc()
}
\value{
TODO
}
\description{
install pandoc if not already installed on the system
}
\section{Version}{
  $Id: ab696f28eb22651bc1bfc8772d53d777927e82db $
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

