% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{statistical_mode}
\alias{statistical_mode}
\title{Statistical Mode}
\usage{
statistical_mode(x, na.rm = TRUE)
}
\arguments{
\item{x}{Can be a list, vector or factor.}

\item{na.rm}{Indicates if NA values should be removed before calculation.}
}
\value{
The statistical mode of x. One or multiple values.
}
\description{
Returns the most frequently occuring value in x. If there are multiple values
that occur most frequently this function returns all of them.
}
\details{
Task 2b
}
