% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{visualisation_categorial_variable}
\alias{visualisation_categorial_variable}
\title{Visualize Categorical Variables Relation to Survival on Titanic}
\usage{
visualisation_categorial_variable(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
Returns a grid of six plots visualizing the survival rates by
Passenger Class, Sex, and Embarkation Point.
}
\description{
This function creates visualizations for the relationship between
survival and three categorical variables:
Passenger Class (Pclass), Sex, and Embarkation Point (Embarked)
}
\details{
Task 2a v.
}
