% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_documentation.r
\name{create_documentation}
\alias{create_documentation}
\title{extract roxygen2-style and markdown comments from a single R code file}
\usage{
create_documentation(file_name, markdown = TRUE, roxygen = TRUE, ...)
}
\arguments{
\item{file_name}{name your R code file to be documented.}

\item{markdown}{set to FALSE if you do not want to parse markdown comments
in your file.}

\item{roxygen}{set to FALSE if you do not want to roxygenize your file.}

\item{...}{arguments passed to \code{\link{create_roxygen_documentation}} or
\code{\link{create_markdown_documentation}}.}
}
\value{
a logical vector indicating whether markdown comments and roxygen
annotations were parsed.
}
\description{
this is a wrapper around \code{\link{create_roxygen_documentation}} and
 \code{\link{create_markdown_documentation}}.
}
\section{Version}{
 $Id: 9bbb752b06d887f2115e37c3e9dadd89e40c49c7 $
}
\examples{
create_template(file_name = "my_r_file.r", type = "roxygen_markdown")
create_documentation("my_r_file.r", overwrite = TRUE)
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

